ai-backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
pillow
httpx
torch
ultralytics
